# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Omnisense Market Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  message: string;
}

export interface GenerateIntelligenceBody {
  /** Name of the target company */
  companyName: string;
  /** One-line description of the company's industry/category */
  category: string;
}

export interface CompetitorItem {
  name: string;
  strengths: string;
  gaps: string;
}

export interface DecisionMakerRole {
  title: string;
  rationale: string;
}

export interface ContactRecord {
  name: string;
  title: string;
  email?: string;
  phone?: string;
  linkedin?: string;
  verified: boolean;
}

export interface OutreachDraft {
  linkedinMessage: string;
  emailSubject: string;
  emailBody: string;
  trackingPixelHtml: string;
  trackingLogicExplanation: string;
}

export interface IntelligenceReport {
  companyName: string;
  category: string;
  companyOverview: string;
  marketPosition: string;
  competitorMapping: CompetitorItem[];
  brandActivity: string;
  experientialFootprint: string;
  strategicWatchouts: string;
  decisionMakerRoles: DecisionMakerRole[];
  contactIntelligence: ContactRecord[];
  outreach: OutreachDraft;
  generatedAt: string;
}

export interface ReportSummary {
  id: number;
  companyName: string;
  category: string;
  createdAt: string;
}

export interface ReportDetail {
  id: number;
  companyName: string;
  category: string;
  report: IntelligenceReport;
  createdAt: string;
}