# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Omnisense Market Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Triggers the multi-agent workflow to generate a full intelligence report for a company
 * @summary Generate market intelligence report
 */
export const GenerateIntelligenceBody = zod.object({
  companyName: zod.string().describe("Name of the target company"),
  category: zod
    .string()
    .describe("One-line description of the company's industry\/category"),
});

export const GenerateIntelligenceResponse = zod.object({
  companyName: zod.string(),
  category: zod.string(),
  companyOverview: zod.string(),
  marketPosition: zod.string(),
  competitorMapping: zod.array(
    zod.object({
      name: zod.string(),
      strengths: zod.string(),
      gaps: zod.string(),
    }),
  ),
  brandActivity: zod.string(),
  experientialFootprint: zod.string(),
  strategicWatchouts: zod.string(),
  decisionMakerRoles: zod.array(
    zod.object({
      title: zod.string(),
      rationale: zod.string(),
    }),
  ),
  contactIntelligence: zod.array(
    zod.object({
      name: zod.string(),
      title: zod.string(),
      email: zod.string().optional(),
      phone: zod.string().optional(),
      linkedin: zod.string().optional(),
      verified: zod.boolean(),
    }),
  ),
  outreach: zod.object({
    linkedinMessage: zod.string(),
    emailSubject: zod.string(),
    emailBody: zod.string(),
    trackingPixelHtml: zod.string(),
    trackingLogicExplanation: zod.string(),
  }),
  generatedAt: zod.coerce.date(),
});

/**
 * Returns all previously generated intelligence reports
 * @summary List saved intelligence reports
 */
export const ListReportsResponseItem = zod.object({
  id: zod.number(),
  companyName: zod.string(),
  category: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListReportsResponse = zod.array(ListReportsResponseItem);

/**
 * Returns the full intelligence report by ID
 * @summary Get a specific intelligence report
 */
export const GetReportParams = zod.object({
  id: zod.coerce.number(),
});

export const GetReportResponse = zod.object({
  id: zod.number(),
  companyName: zod.string(),
  category: zod.string(),
  report: zod.object({
    companyName: zod.string(),
    category: zod.string(),
    companyOverview: zod.string(),
    marketPosition: zod.string(),
    competitorMapping: zod.array(
      zod.object({
        name: zod.string(),
        strengths: zod.string(),
        gaps: zod.string(),
      }),
    ),
    brandActivity: zod.string(),
    experientialFootprint: zod.string(),
    strategicWatchouts: zod.string(),
    decisionMakerRoles: zod.array(
      zod.object({
        title: zod.string(),
        rationale: zod.string(),
      }),
    ),
    contactIntelligence: zod.array(
      zod.object({
        name: zod.string(),
        title: zod.string(),
        email: zod.string().optional(),
        phone: zod.string().optional(),
        linkedin: zod.string().optional(),
        verified: zod.boolean(),
      }),
    ),
    outreach: zod.object({
      linkedinMessage: zod.string(),
      emailSubject: zod.string(),
      emailBody: zod.string(),
      trackingPixelHtml: zod.string(),
      trackingLogicExplanation: zod.string(),
    }),
    generatedAt: zod.coerce.date(),
  }),
  createdAt: zod.coerce.date(),
});