# lib/api-zod/src/generated/types/generateIntelligenceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Omnisense Market Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateIntelligenceBody {
  /** Name of the target company */
  companyName: string;
  /** One-line description of the company's industry/category */
  category: string;
}