# lib/api-zod/src/generated/types/intelligenceReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Omnisense Market Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { CompetitorItem } from "./competitorItem";
import type { ContactRecord } from "./contactRecord";
import type { DecisionMakerRole } from "./decisionMakerRole";
import type { OutreachDraft } from "./outreachDraft";

export interface IntelligenceReport {
  companyName: string;
  category: string;
  companyOverview: string;
  marketPosition: string;
  competitorMapping: CompetitorItem[];
  brandActivity: string;
  experientialFootprint: string;
  strategicWatchouts: string;
  decisionMakerRoles: DecisionMakerRole[];
  contactIntelligence: ContactRecord[];
  outreach: OutreachDraft;
  generatedAt: Date;
}