# lib/api-zod/src/generated/types/outreachDraft.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Omnisense Market Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface OutreachDraft {
  linkedinMessage: string;
  emailSubject: string;
  emailBody: string;
  trackingPixelHtml: string;
  trackingLogicExplanation: string;
}