# lib/api-zod/src/generated/types/reportDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Omnisense Market Intelligence Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { IntelligenceReport } from "./intelligenceReport";

export interface ReportDetail {
  id: number;
  companyName: string;
  category: string;
  report: IntelligenceReport;
  createdAt: Date;
}